fix the profile URL used for auth logins without a username.
