+++ app/js/utils/profile-utils.js
@@ -3,6 +3,10 @@ import { decodeToken, TokenVerifier } from 'jsontokens'
 
 import ecurve from 'ecurve'
 import { ECPair as ECKeyPair } from 'bitcoinjs-lib'
+import log4js from 'log4js'
+
+const logger = log4js.getLogger('utils/profile-utils.js')
+
 const secp256k1 = ecurve.getCurveByName('secp256k1')
 
 export function verifyToken(token, verifyingKeyOrAddress) {
@@ -98,6 +102,64 @@ export function getProfileFromTokens(tokenRecords, publicKeychain, silentVerify 
   })
 
   return profile
+}
+
+/**
+ * Try to fetch and verify a profile from the historic set of default locations,
+ * in order of recency. If all of them return 404s, or fail to validate, return null
+ */
+export function fetchProfileLocations(gaiaUrlBase: string,
+                                      ownerAddress: string,
+                                      firstAddress: string,
+                                      ownerIndex: number) {
+  function recursiveTryFetch(locations) {
+    if (locations.length === 0) {
+      return Promise.resolve(null)
+    }
+    const location = locations[0]
+    return fetch(location)
+      .then(response => {
+        if (response.ok) {
+          return response.json()
+            .then(tokenRecords => getProfileFromTokens(tokenRecords, ownerAddress, false))
+            .then(profile => {
+              logger.debug(`Found valid profile at ${location}`)
+              return { profile, profileUrl: location }
+            })
+            .catch(() => {
+              logger.debug(`Failed to verify profile at ${location}... trying others`)
+              return recursiveTryFetch(locations.slice(1))
+            })
+        } else {
+          logger.debug(`Failed to find profile at ${location}... trying others`)
+          return recursiveTryFetch(locations.slice(1))
+        }
+      })
+      .catch(() => {
+        logger.debug(`Error in fetching profile at ${location}... trying others`)
+        return recursiveTryFetch(locations.slice(1))
+      })
+  }
+
+  const urls = []
+  // the new default
+  urls.push(`${gaiaUrlBase}/${ownerAddress}/profile.json`)
+
+  // the 'indexed' URL --
+  //  this is gaia/:firstAddress/:index/profile.json
+  //  however, the index is _not_ equal to the current index.
+  //  indexes were mapped from
+  //    correct: [0, 1, 3, 5, 7, 9...]
+  //  incorrect: [0, 1, 2, 3, 4, 5...]
+
+  if (ownerIndex < 2) {
+    urls.push(`${gaiaUrlBase}/${firstAddress}/${ownerIndex}/profile.json`)
+  } else if (ownerIndex % 2 === 1) {
+    const buggedIndex = 1 + Math.floor(ownerIndex / 2)
+    urls.push(`${gaiaUrlBase}/${firstAddress}/${buggedIndex}/profile.json`)
+  }
+
+  return recursiveTryFetch(urls)
 }
 
 export function signProfileForUpload(profile, keypair) {
